Specify the supported hashing algorithms. (#126)
Given that we haven't changed the supported algorithms in almost a decade, it
seems reasonable to lock them in and define their ordering rather than
deferring to the user agent.

If we add new algorithms in the future, this specificity will make it
simpler to reason about how those algorithms fit into the system.

This is not strictly an editorial change, as it does introduce new
requirements for user agents. Practically, however, it only encodes the
status quo, and will not result in any change in existing tests or
behavior.

Also: this closes w3c/webappsec-subresource-integrity#106 via an adjusted
note in a section the patch was touching anyway.